:sparkles: add ubication for IP  - :hankey: The ip ubication found it in reconquista but i am in Madryn

+++ app/src/main/java/com/example/saba/weatherapp/services/WeatherService.kt
@@ -12,7 +12,7 @@ import retrofit2.http.Query
 interface WeatherService {
 
     @GET("/data/2.5/weather")
-    fun getWeather(@Query("lat") lat:Long, @Query("lon") lon:Long) : Call<Weather>
+    fun getWeather(@Query("lat") lat:Double, @Query("lon") lon:Double) : Call<Weather>
 
     @GET("img/w/{icon}.png")
     fun getIcon(@Path("icon") icon: String?) : Call<Image>
